+++ src/adzerk/boot_cljs.clj
@@ -300,3 +300,24 @@
               (core/add-meta (cljs-output-meta tmp-result))
               (core/add-meta (deps/compiled fileset dep-order))
               core/commit!))))))
+
+(core/deftask cljs-edn
+  "Generate a .cljs.edn file."
+  [e edn      VAL str   "EDN file name."
+   r require  VAL [sym] "Vector of namespaces to require."
+   f init-fns VAL [sym] "Vector of fuctions to wrap in do block."
+   o compiler-options VAL {} "A map of compiler options."]
+   (assert (:edn *opts*) "Must provide an edn file name.")
+   (assert (:init-fns *opts*) "Must provide an init-fn.")
+   (let [edn    (:edn *opts*)
+         init   (:init-fn *opts*)
+         ednstr {:require (:require *opts* [])
+                 :init-fns (:init-fns *opts* [])
+                 :compiler-options (:compiler-options *opts* {})}
+         tmp    (core/tmp-dir!)
+         fname  (str edn ".cljs.edn")
+         fedn   (io/file tmp fname)]
+     (core/with-pre-wrap fileset
+       (info (str "Generating EDN file: " fname "\n"))
+       (doto fedn (spit ednstr))
+       (-> fileset (core/add-resource tmp) core/commit!))))