+++ CHANGES.md
@@ -27,6 +27,7 @@
 which included change [CLJS-1367](https://github.com/clojure/clojurescript/commit/9e662c3ea5b9f536f303e9084415e988bf5d0878), should work.
 - Expose ClojureScript compiler options in `.cljs.edn` Boot fileset metadata for
 others tasks (e.g. boot-reload)
+- Task to generate `.cljs.edn` file during build, output to fileset.
 - Hide stacktraces for Cljs compiler exceptions (reader exception, analysis error)
 
 **[compare](https://github.com/adzerk-oss/boot-cljs/compare/1.7.228-2...2.0.0)**